Remove duplicate stub files and consolidate to production location
- Delete adgn/src/adgn/mcp/testing/{typed_stubs.py,server_stubs.py}
- Update imports in test and stub files to use adgn.mcp.stubs instead
- Add StructuredContent type alias to client_helpers.py
- Add Usage alias in model.py for backwards compatibility with tests

All stub utilities now live in adgn/src/adgn/mcp/stubs/ only.
